requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
aiosqlite==0.19.0  # For development

# Caching & Queue
redis==5.0.1
celery==5.3.4

# Workflow
temporalio==1.0.0

# AI/ML
openai==1.3.0
anthropic==0.7.7
tiktoken==0.5.1

# Scraping
playwright==1.40.0
httpx==0.25.1
beautifulsoup4==4.12.2
feedparser==6.0.10

# Email
aiosmtplib==2.0.2
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0

# Storage
aioboto3
boto3

# File parsing
python-docx==1.1.0
PyPDF2==3.0.1
markdown==3.5.1

# Utilities
python-dateutil==2.8.2
tenacity==8.2.3
jinja2==3.1.2
pyyaml==6.0.1






pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
factory-boy==3.3.0
faker==20.1.0

# Code quality
black==23.11.0
isort==5.12.0
ruff==0.1.6
mypy==1.7.0
pre-commit==3.5.0

# Debugging
ipython==8.17.2
python-multipart==0.0.6
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
watchfiles==0.21.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.14